Make Kubernetes resources Hashable
The following resources:

- apiextensions.v1.CustomResourceValidation
- apps.v1.ControllerRevision
- meta.v1.ManagedFieldsEntry
- meta.v1.WatchEvent

contains fields that are arbitrary JSON objects, which were represented
as a Dictionary<String, Any> up until now. This prevented equality and
hashability of those resources.

A new simple JSONObject type is introduced, which acts a simple wrapper
around the dictionary.

The implementation assumes a correct JSON object and uses a forced
JSON encoding in order to convert the dictionary to a Data instance,
which in turn is used for equality and hashing.